Add backup script to scan for donation transactions and index them
